mp3: Try partial decode on main_data underflow.
If the bit resevoir does not have enough bits to fully read the
main data of both granules, drop the granule(s) associated with
the missing bits. In many cases there will be enough data to
decode the second granule.

Previously, the entire packet would not be decoded, but this
improvement will allow more audio to be preserved in the case of
an error.